Sync the gigabeat branch with trunk. Also a few small changes that allow the main binary to be compiled.
git-svn-id: svn://svn.rockbox.org/rockbox/branches/gigabeat-s@14400 a1c6a512-1295-4272-9138-f99709370657

+++ rbutil/ipodpatcher/ipodio-posix.c
@@ -143,12 +143,12 @@ int ipod_seek(struct ipod_t* ipod, unsigned long pos)
     return 0;
 }
 
-int ipod_read(struct ipod_t* ipod, unsigned char* buf, int nbytes)
+ssize_t ipod_read(struct ipod_t* ipod, unsigned char* buf, int nbytes)
 {
     return read(ipod->dh, buf, nbytes);
 }
 
-int ipod_write(struct ipod_t* ipod, unsigned char* buf, int nbytes)
+ssize_t ipod_write(struct ipod_t* ipod, unsigned char* buf, int nbytes)
 {
     return write(ipod->dh, buf, nbytes);
 }
