+++ rbutil/ipodpatcher/ipodio-win32.c
@@ -173,7 +173,7 @@ int ipod_seek(struct ipod_t* ipod, unsigned long pos)
     return 0;
 }
 
-int ipod_read(struct ipod_t* ipod, unsigned char* buf, int nbytes)
+ssize_t ipod_read(struct ipod_t* ipod, unsigned char* buf, int nbytes)
 {
     unsigned long count;
 
@@ -185,7 +185,7 @@ int ipod_read(struct ipod_t* ipod, unsigned char* buf, int nbytes)
     return count;
 }
 
-int ipod_write(struct ipod_t* ipod, unsigned char* buf, int nbytes)
+ssize_t ipod_write(struct ipod_t* ipod, unsigned char* buf, int nbytes)
 {
     unsigned long count;
